Added message templates for invitation responses
