+++ contentcuration/contentcuration/api.py
@@ -173,3 +173,16 @@ def batch_add_tags(request):
     ThroughModel.objects.bulk_create(bulk_list)
 
     return HttpResponse("Tags are successfully saved.", status=200)
+
+
+def add_editor_to_channel(invitation):
+    if invitation.share_mode == "view":
+        if invitation.invited in invitation.channel.editors.all():
+            invitation.channel.editors.remove(invitation.invited)
+        invitation.channel.viewers.add(invitation.invited)
+    else:
+        if invitation.invited in invitation.channel.viewers.all():
+            invitation.channel.viewers.remove(invitation.invited)
+        invitation.channel.editors.add(invitation.invited)
+    invitation.channel.save()
+    invitation.delete()
